% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_uniprot}
\alias{anno_uniprot}
\title{Data.table with uniprot description}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 41748 rows and 2 columns.
}
\usage{
data(anno_uniprot)
}
\description{
Annotation matrix Uniprot and related description
}
\keyword{datasets}
