% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correction}
\alias{batch_correction}
\title{Apply Batch Effect Correction to Proteome Data}
\usage{
batch_correction(proteome_data, batch_col = "batch")
}
\arguments{
\item{proteome_data}{A list containing proteome data and sample annotations.}

\item{batch_col}{A string indicating the column name in the sample annotation to be used for batch correction (default is "batch").}
}
\value{
A list containing the batch-corrected proteome data, with \code{dat_pep} and \code{dat_gene} corrected for batch effects.
}
\description{
This function applies batch effect correction to proteome data using the \code{proBatch} package.
It adjusts both peptide and gene data by correcting for batch effects based on the specified batch column.
}
\details{
The function uses the ComBat method for batch effect correction.
It adjusts both peptide and gene data separately, based on the sample annotation provided.
}
\examples{
\dontrun{
corrected_data <- batch_correction(proteome_data = proteome_data, batch_col = "batch")
}

}
