% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{correlation Heatmap of Selected Proteins}
\usage{
correlation_heatmap(
  proteome_data,
  replicate,
  mode = "correlation",
  cov_thr = 0.9999,
  protein_list = NULL,
  algorithm = "pearson"
)
}
\arguments{
\item{proteome_data}{A list containing proteome data, including gene intensity data and sample annotations.}

\item{replicate}{Character. A Sample or a Condition to represent in the heatmap.}

\item{mode}{Character. Mode of proteins selection. Choose between: \code{correlation}: show only protein with correlation > thr (DEFAULT); \code{protein_list}: show only protein provided in input.}

\item{cov_thr}{Double. Numeric threshold of correlation. (Default: 0.9)}

\item{protein_list}{A character vector of gene names to be plotted.}

\item{algorithm}{Character. Pearson or Spearman (Default: Pearson).}
}
\value{
A list containing:
\item{dt}{Data table of protein intensity values and the corresponding sample and condition information.}
\item{plot}{ggplot2 object representing the protein abundance plot for the selected proteins.}
}
\description{
This function generates a ...
}
\examples{
\dontrun{
result <- correlation_heatmap(proteome_data, list_protein = c("ProteinA", "ProteinB"))
print(result$plot)
}

}
