% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprocess_phosphorilation.R
\name{create_phosphosite_plot}
\alias{create_phosphosite_plot}
\title{Create Phosphosite Distribution Plot}
\usage{
create_phosphosite_plot(proteome_data, software)
}
\arguments{
\item{proteome_data}{A list containing \code{psm_peptide_table}, which includes peptide modification information.}

\item{software}{Character. The software used for processing (\code{"PD"} for Proteome Discoverer, \code{"MQ"} for MaxQuant).}
}
\value{
A list containing:
\itemize{
\item \code{dt}: A data table with phosphosite percentages.
\item \code{plot}: A ggplot2 object visualizing the phosphosite distribution.
}
}
\description{
This function generates a lollipop plot showing the distribution of phosphorylated residues (S, T, Y)
in phosphoproteomics data, based on the output of either Proteome Discoverer (PD) or MaxQuant (MQ).
}
\examples{
\dontrun{
  phospho_plot <- create_phosphosite_plot(proteome_data, software="MQ")
}
}
