% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{createFrequencyMat}
\alias{createFrequencyMat}
\title{Create frequency matrix}
\usage{
createFrequencyMat(substrates.seq)
}
\arguments{
\item{substrates.seq}{A substrate sequence}
}
\value{
A frequency matrix of amino acid from substrates.seq.
}
\description{
Create frequency matrix
}
