% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{differential_analysis}
\alias{differential_analysis}
\title{Differential Analysis for Proteomic and Phosphoproteomic Data}
\usage{
differential_analysis(
  proteome_data,
  formule_contrast,
  fc_thr = 0.75,
  pval_fdr = "p_val",
  pval_thr = 0.05,
  signal_thr = 0
)
}
\arguments{
\item{proteome_data}{A list containing proteomic data.}

\item{formule_contrast}{A vector of formulas defining the contrasts for differential expression analysis.}

\item{fc_thr}{Numeric. The threshold for fold change (default is 0.75).}

\item{pval_fdr}{String. The column name for p-value after FDR adjustment (default is "p_val").}

\item{pval_thr}{Numeric. The threshold for p-value (default is 0.05).}

\item{signal_thr}{Numeric. The threshold for signal log2 intensity (default is 0).}
}
\value{
A list containing the results of the differential analysis
}
\description{
This function performs differential expression analysis for proteomic and phosphoproteomic data, using limma for statistical analysis. It handles the combination of proteome and phosphoproteome data, calculates fold change, p-value, and other statistical parameters, and returns results in both long and wide formats.
}
\details{
If the input data includes phosphoproteomic and proteomic data, the function will merge these datasets, perform the differential analysis for both types, and return the results for both proteins and peptides. If only proteomic data is available, the function will analyze proteins and peptides separately.
}
\examples{
\dontrun{
results <- differential_analysis(proteome_data = proteome_data, 
                                 formule_contrast = c("comparison1"="condition1-condition2"),
                                 fc_thr = 0.75, pval_thr = 0.05)
}

}
