% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_figure.R
\name{enrichment_figure}
\alias{enrichment_figure}
\title{Generate Enrichment Plots from Enrichment Analysis Results}
\usage{
enrichment_figure(
  enr_df,
  category = c("all"),
  enrich_filter_term = NULL,
  overlap_size_enrich_thr = 5,
  save = F,
  color_contrast = NULL,
  dirOutput = "results_ProTN",
  subfolder = "pics",
  namefile = "enrichement_plot"
)
}
\arguments{
\item{enr_df}{A data.table containing the enrichment results, typically generated from
the \code{perform_enrichment_analysis} function.}

\item{category}{A vector of categories to filter the enrichment results.
Possible values: "all", "up", "down". Default is \code{c("all")}.}

\item{enrich_filter_term}{A vector of terms to filter the enrichment results. Default is \code{NULL}.
If provided, only enrichment terms containing any of these words will be kept.}

\item{overlap_size_enrich_thr}{The minimum overlap size (number of DEPs) required to retain
an enrichment term. Default is 5.}

\item{save}{A logical flag to indicate if the resulting plots should be saved to disk.
Default is \code{FALSE}. If \code{TRUE}, plots are saved as individual PDFs and combined into one.}

\item{color_contrast}{A vector of colors to be used for the plots. Default is \code{NULL},
in which case a default color scheme is applied.}

\item{dirOutput}{The directory where the results will be saved. Default is \code{"results_ProTN"}.}

\item{subfolder}{The subfolder within \code{dirOutput} where the plot files will be saved.
Default is \code{"pics"}.}

\item{namefile}{The base name for the combined PDF file if \code{save} is \code{TRUE}. Default is \code{"enrichment_plot"}.}
}
\value{
A list of \code{ggplot} objects (one for each category) that represent the enrichment plots
visualizing the significant terms based on the selected categories and filtering criteria.
}
\description{
This function generates a set of plots visualizing the enrichment results obtained
from a differential expression analysis. It allows for filtering the enrichment results
based on specific categories, terms, and overlap size. The plots are then saved as PDF files
for further inspection.
}
\details{
The function supports multiple categories and allows customization of colors for different
groups. The resulting plots are saved as individual PDFs and combined into a single PDF file
if the \code{save} parameter is set to TRUE.
}
\examples{
# Example usage
plotlist <- enrichment_figure(enr_df = enr_results, category = c("all"), save = TRUE)

}
