% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{extract_example}
\alias{extract_example}
\title{Extract case study}
\usage{
extract_example(path_phospho = NULL, path_proteome = NULL)
}
\arguments{
\item{path_phospho}{Character; path where extract case study}

\item{path_proteome}{Character; path where extract case study}
}
\description{
Extract case study
}
\examples{
\dontrun{
extract_example(path_proteome = tempdir())
}
}
