% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{frequencyScoring}
\alias{frequencyScoring}
\title{Frequency scoring}
\usage{
frequencyScoring(sequence.list, frequency.mat)
}
\arguments{
\item{sequence.list}{A vector list of sequences}

\item{frequency.mat}{A matrix output from \code{createFrequencyMat}}
}
\value{
A vector of frequency score
}
\description{
Frequency scoring
}
