% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{generate_abundance_plate_plot}
\alias{generate_abundance_plate_plot}
\title{Generate Abundance Plot for the plate of Spatial Proteomics}
\usage{
generate_abundance_plate_plot(proteome_data)
}
\arguments{
\item{proteome_data}{A list containing \code{c_anno} or \code{c_anno_phospho}, \code{psm_log_prot_df}, and \code{psm_log_pet_df}.}
}
\value{
A list containing:
\itemize{
\item \code{dt}: Data table for abundance.
\item \code{plot}: A ggplot2 object visualizing abundance.
}
}
\description{
This function generates an abundance plot.
}
\examples{
\dontrun{
  abundance_plot <- generate_abundance_plate_plot(proteome_data)
}
}
