% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{generate_abundance_plot}
\alias{generate_abundance_plot}
\title{Generate Abundance Plot}
\usage{
generate_abundance_plot(proteome_data, phospho_with_proteome = FALSE)
}
\arguments{
\item{proteome_data}{A list containing \code{c_anno} or \code{c_anno_phospho}, \code{psm_log_prot_df}, and \code{psm_log_pet_df}.}

\item{phospho_with_proteome}{Logical. Whether to include phosphoproteome data alongside proteome data (default: \code{FALSE}).}
}
\value{
A list containing:
\itemize{
\item \code{dt}: Data table for abundance.
\item \code{plot}: A ggplot2 object visualizing abundance.
}
}
\description{
This function generates an abundance plot.
}
\examples{
\dontrun{
  abundance_plot <- generate_abundance_plot(proteome_data)
}
}
