% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{generate_differential_barplots}
\alias{generate_differential_barplots}
\title{Generate Differential Barplots}
\usage{
generate_differential_barplots(
  differential_results,
  data_type = "protein",
  color_contrast = NULL,
  phospho_ctrl = FALSE
)
}
\arguments{
\item{differential_results}{A list containing the differential analysis results.}

\item{data_type}{A character string indicating the type of data to plot. Must be either "protein" or "peptide".}

\item{color_contrast}{A vector of colors to use for the plot (optional).}

\item{phospho_ctrl}{A logical value indicating whether to include phospho-control comparisons in the plot (default is FALSE).}
}
\value{
A list containing the lollipop plot for differential expression results.
}
\description{
This function generates differential barplots based on the differential analysis results.
It creates a lollipop plot to visualize the comparison of protein or peptide differential expression results.
}
\details{
This function filters the differential results based on the specified parameters,
such as whether to include phospho-control comparisons or not, and generates a
lollipop plot to visualize the differential expression of proteins or peptides.
}
\examples{
\dontrun{
plot_results <- generate_differential_barplots(differential_results = differential_results,
                                               data_type = "protein")
}

}
