% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{generate_peptide_distribution_plot}
\alias{generate_peptide_distribution_plot}
\title{Generate Peptide Distribution Plot}
\usage{
generate_peptide_distribution_plot(
  proteome_data,
  phospho_with_proteome = FALSE
)
}
\arguments{
\item{proteome_data}{A list containing proteome data, including peptide annotation tables.}

\item{phospho_with_proteome}{Logical; whether the dataset contains phosphoproteomics data alongside proteomics data. Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{dt}: Data table for abundance distribution.
\item \code{plot}: A ggplot2 object visualizing abundance distribution.
}
}
\description{
This function generates a peptide distribution plot for a given proteome dataset.
}
\examples{
\dontrun{
result <- generate_peptide_distribution_plot(proteome_data)
}
}
