% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{generate_upset_plot}
\alias{generate_upset_plot}
\title{Generate an UpSet plot for differential results}
\usage{
generate_upset_plot(
  differential_results,
  type = "protein",
  DE_class = "all",
  nintersects = 20,
  remove_zero = TRUE,
  order.intersect.by = "size",
  order.set.by = "size",
  relative_height = 3,
  relative_width = 0.3,
  top.bar.color = "#7AC2E0DD",
  top.bar.y.label = NULL,
  top.bar.show.numbers = TRUE,
  top.bar.numbers.size = 3,
  sets.bar.color = "#7f7f7fAA",
  sets.bar.show.numbers = FALSE,
  sets.bar.x.label = "Set Size",
  intersection.matrix.color = "#7AC2E0DD",
  specific = TRUE,
  ...
)
}
\arguments{
\item{differential_results}{A list containing peptide_results_long data.table.}

\item{type}{Character; specifies the type of data to plot. Options are "protein" or "peptide".}

\item{DE_class}{Character, one of "all", "up", "down". Specifies which class to plot.}

\item{nintersects}{Integer, number of intersections to display.}

\item{remove_zero}{Logical, If TRUE (default) remove intersection with 0 proteins.}

\item{order.intersect.by}{Character, order intersections by "size", "name", or "none".}

\item{order.set.by}{Character, order sets by "size", "name", or "none".}

\item{relative_height}{Numeric, relative height of the plot.}

\item{relative_width}{Numeric, relative width of the plot.}

\item{top.bar.color}{Character, color for the top bar.}

\item{top.bar.y.label}{Character, label for the top bar y-axis.}

\item{top.bar.show.numbers}{Logical, show numbers on top bar.}

\item{top.bar.numbers.size}{Numeric, size of numbers on top bar.}

\item{sets.bar.color}{Character, color for the sets bar.}

\item{sets.bar.show.numbers}{Logical, show numbers on sets bar.}

\item{sets.bar.x.label}{Character, label for the sets bar x-axis.}

\item{intersection.matrix.color}{Character, color for the intersection matrix.}

\item{specific}{Logical, show specific intersections.}

\item{...}{Additional arguments passed to ggVennDiagram::plot_upset.}
}
\value{
A list containing:
\item{list}{List containing the gene names of each comparison.}
\item{plot}{ggplot2 object representing the MDS plot with sample labels and color-coded conditions.}
}
\description{
Generate an UpSet plot for differential results
}
\examples{
\dontrun{
result <- generate_upset_plot(differential_results, type)
print(result$plot)
}

}
