% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{generate_volcano_plots}
\alias{generate_volcano_plots}
\title{Generate Volcano Plots for Differential Expression Analysis}
\usage{
generate_volcano_plots(
  differential_results,
  data_type = NULL,
  comparison = NULL,
  fc_thr = 0.75,
  pval_fdr = "p_val",
  pval_thr = 0.05,
  color_contrast = NULL
)
}
\arguments{
\item{differential_results}{A list containing the differential analysis results.}

\item{data_type}{A character string indicating the type of data to plot. Must be either "protein" or "peptide" (default is NULL).}

\item{comparison}{A character string specifying the comparison to plot (e.g., "comparison1").}

\item{fc_thr}{A numeric value for the fold-change threshold (default is 0.75).}

\item{pval_fdr}{A character string indicating whether to use "p_val" or "p_adj" for the p-value threshold (default is "p_val").}

\item{pval_thr}{A numeric value for the p-value threshold (default is 0.05).}

\item{color_contrast}{A vector of two colors to use for plotting (optional).}
}
\value{
A list containing the volcano plot for the specified comparison.
}
\description{
This function generates volcano plots for visualizing the differential expression of proteins or peptides.
It plots the log2 fold change (log2_FC) against the -log10 p-value (or FDR adjusted p-value),
highlighting significant changes based on fold-change and p-value thresholds.
}
\details{
This function generates volcano plots to visualize differential expression analysis,
with fold-change and p-value (or FDR adjusted p-value) on the axes.
It highlights genes/peptides based on the fold-change and p-value thresholds.
}
\examples{
# Example usage:
volcano_plot <- generate_volcano_plots(differential_results = differential_results,
                                       comparison = "comparison1",
                                       fc_thr = 1, pval_thr = 0.05)

}
