% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{getSPS}
\alias{getSPS}
\title{Generate set of stable phosphoporylated sites}
\usage{
getSPS(phosData, assays, conds, num)
}
\arguments{
\item{phosData}{a list of users' PhosphoExperiment objects from which
generate SPSs}

\item{assays}{an assay to use for each dataset in phosData}

\item{conds}{a list of vector contains the conditions labels for each sample
in the phosphoExperiment objects}

\item{num}{the number of identified SPSs, by default is 100}
}
\value{
A vectors of stably phosphorylated sites
}
\description{
Generate set of stable phosphoporylated sites
}
