% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_matrix}
\alias{impute_matrix}
\title{Matrix Imputation Function}
\usage{
impute_matrix(mat, c_anno, type)
}
\arguments{
\item{mat}{Data frame or matrix; the intensity matrix to be imputed. The first column must be a peptide ID column, and the rest should contain intensity values.}

\item{c_anno}{Data frame; a data frame containing sample annotations, with columns for the sample ID and condition.}

\item{type}{character; Type of imputation selected (\code{phosr}, \code{gaussian}, \code{missForest}, \code{pcaMethods}).}
}
\value{
A data.table with the imputed intensity matrix, retaining the original peptide IDs.
}
\description{
This function imputes missing values in a matrix using either PhosR's imputation method (scImpute + tImpute) for data or a median imputation approach if PhosR fails.
}
