% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinase_tree.R
\name{kinase_tree}
\alias{kinase_tree}
\title{Kinase Activity Tree Calculation}
\usage{
kinase_tree(
  proteome_data,
  differential_results,
  formule_CORAL,
  dirOutput = "results_ProTN",
  subfold = "pics",
  phospho_ctrl = FALSE
)
}
\arguments{
\item{proteome_data}{List; a list containing proteomics data.}

\item{differential_results}{List; a list containing differential analysis results, specifically \code{peptide_results_long} for peptide-wise differential analysis.}

\item{formule_CORAL}{List; a list of formulae for CORAL-based kinase activity calculations. It must include formulae for the comparisons of interest.}

\item{dirOutput}{Character; the directory where the results will be stored. Default is "results_ProTN".}

\item{subfold}{Character; the subdirectory within \code{dirOutput} to store the results, default is "pics".}

\item{phospho_ctrl}{Logical; whether to include phospho-control data in the calculations. Default is FALSE (excludes phospho-control data).}
}
\value{
A list of SVG plots, one for each comparison in the \code{formule_CORAL} list, showing the kinase activity tree for each comparison.
}
\description{
This function calculates kinase activity for differential analysis results, creating a tree-based visualization of kinase activity. It uses CORAL-based formulae and generates kinase activity plots. The function allows the user to specify whether to include phospho-control data or not.
}
\examples{
\dontrun{
kinase_tree_results <- kinase_tree(proteome_data = proteome_data, 
                                   differential_results = differential_results, 
                                   formule_CORAL = formule_CORAL)
}

}
