% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{kinaseSubstrateHeatmap_local}
\alias{kinaseSubstrateHeatmap_local}
\title{Kinase-substrate annotation prioritisation heatmap}
\usage{
kinaseSubstrateHeatmap_local(
  phosScoringMatrices,
  top = 3,
  printPlot = NULL,
  filePath = "./kinaseSubstrateHeatmap.pdf",
  width = 20,
  height = 20
)
}
\arguments{
\item{phosScoringMatrices}{a matrix returned from kinaseSubstrateScore_local}

\item{top}{the number of top ranked phosphosites for each kinase to be
included in the heatmap. Default is 1.}

\item{printPlot}{indicate whether the plot should be saved as a PDF
in the specified directory. Default is NULL, otherwise specify TRUE.}

\item{filePath}{path name to save the plot as a PDF file.
Default saves in the working directory.}

\item{width}{width of PDF.}

\item{height}{height of PDF.}
}
\value{
a pheatmap object.
}
\description{
Kinase-substrate annotation prioritisation heatmap
}
