% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{kinaseSubstratePred}
\alias{kinaseSubstratePred}
\title{kinaseSubstratePred}
\usage{
kinaseSubstratePred(
    phosScoringMatrices,
    ensembleSize = 10,
    top = 50,
    cs = 0.8,
    inclusion = 20,
    iter = 5,
    verbose = TRUE
)
}
\arguments{
\item{phosScoringMatrices}{An output of kinaseSubstrateScore_local}

\item{ensembleSize}{An ensemble size.}

\item{top}{a number to select top kinase substrates.}

\item{cs}{Score threshold.}

\item{inclusion}{A minimal number of substrates required for a kinase to be
selected.}

\item{iter}{A number of iterations for adaSampling.}

\item{verbose}{Default to \code{TRUE} to show messages during the progress.
All messages will be suppressed if set to \code{FALSE}}
}
\value{
Kinase prediction matrix
}
\description{
A machine learning approach for predicting specific kinase for a given
substrate. This prediction framework utilise adaptive sampling.
}
