% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{kinaseSubstrateProfile}
\alias{kinaseSubstrateProfile}
\title{Kinase substrate profiling}
\usage{
kinaseSubstrateProfile(substrate.list, mat)
}
\arguments{
\item{substrate.list}{a list of kinases with each element containing an array
of substrates.}

\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to samples.}
}
\value{
Kinase profile list.
}
\description{
This function generates substrate profiles for kinases that have
one or more substrates quantified in the phosphoproteome data.
}
