% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{kinaseSubstrateScore_local}
\alias{kinaseSubstrateScore_local}
\title{Kinase substrate scoring}
\usage{
kinaseSubstrateScore_local(
  substrate.list,
  comp,
  mat,
  seqs,
  numMotif = 5,
  numSub = 1,
  species = "mouse",
  verbose = TRUE,
  file_heatmap = "./heatmap.pdf"
)
}
\arguments{
\item{substrate.list}{A list of kinases with each element containing an array
of substrates.}

\item{comp}{Comparison}

\item{mat}{A matrix with rows correspond to phosphosites and columns
correspond to samples.}

\item{seqs}{An array containing aa sequences surrounding each of all
phosphosites.
Each sequence has length of 15 (-7, p, +7).}

\item{numMotif}{Minimum number of sequences used for compiling motif for
each kinase.
Default is 5.}

\item{numSub}{Minimum number of phosphosites used for compiling
phosphorylation
profile for each kinase. Default is 1.}

\item{species}{Motif list species to be used. Currently there are
\code{mouse} (default), \code{human} and \code{rat}.}

\item{verbose}{Default to \code{TRUE} to show messages during the progress.
All messages will be suppressed if set to \code{FALSE}}

\item{file_heatmap}{Path and file name for the heatmap}
}
\value{
A list of 4 elements.
\code{motifScoreMatrix}, \code{profileScoreMatrix},
\code{combinedScoreMatrix}, \code{ksActivityMatrix} (kinase activity matrix)
and their \code{weights}.
}
\description{
This function generates substrate scores for kinases that pass
filtering based on both motifs and dynamic profiles
}
