% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{ma_plot}
\alias{ma_plot}
\title{Abundance vs Fold Change}
\usage{
ma_plot(
  differential_results,
  proteome_data,
  type,
  condition,
  comparison,
  col_vec = NULL
)
}
\arguments{
\item{differential_results}{A list containing the differential analysis results.}

\item{proteome_data}{A list containing proteome data, including sample annotation and either protein or peptide data matrices.}

\item{type}{Character; specifies the type of data to plot. Options are "protein" or "peptide".}

\item{condition}{Character; Specify the condition or conditions to represent.}

\item{comparison}{Character; Specify the comparison to represent.}

\item{col_vec}{Character. Vector with 3 colors.}
}
\value{
A list containing:
\item{dt}{Data table containing the coordinates for each sample.}
\item{plot}{ggplot2 object representing the MDS plot with sample labels and color-coded conditions.}
}
\description{
This function generates ...
}
\examples{
\dontrun{
result <- ma_plot(differential_results, proteome_data, type, condition, comparison)
print(result$plot)
}

}
