% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{matANOVA}
\alias{matANOVA}
\title{ANOVA test}
\usage{
matANOVA(mat, grps)
}
\arguments{
\item{mat}{An p by n matrix where p is the number of phosphosites and n is
the number of samples}

\item{grps}{A vector of length n, with group or time point information of the
samples}
}
\value{
A vector of multiple testing adjusted p-values
}
\description{
Performs an ANOVA test and returns its adjusted p-value
}
