% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{mds_differential_analysis_plot}
\alias{mds_differential_analysis_plot}
\title{MDS Plot}
\usage{
mds_differential_analysis_plot(differential_analysis, proteome_data, type)
}
\arguments{
\item{differential_analysis}{A list containing the differential analysis results.}

\item{proteome_data}{A list containing proteome data, including sample annotation and either protein or peptide data matrices.}

\item{type}{Character; specifies the type of data to plot. Options are "protein" or "peptide".}
}
\value{
A list containing:
\item{mds_dt}{Data table containing the MDS coordinates for each sample.}
\item{plot}{ggplot2 object representing the MDS plot with sample labels and color-coded conditions.}
}
\description{
This function generates a Multi-Dimensional Scaling (MDS) plot for either protein or peptide data, based on a Euclidean distance matrix of sample similarities.
It supports the inclusion of phosphoproteomics data if available.
}
\examples{
\dontrun{
result <- mds_plot(proteome_data, type = "protein")
print(result$plot)
}

}
