% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{meanAbundance}
\alias{meanAbundance}
\title{Obtain average expression from replicates}
\usage{
meanAbundance(mat, grps)
}
\arguments{
\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to samples.}

\item{grps}{a string specifying the grouping (replciates).}
}
\value{
a matrix with mean expression from replicates
}
\description{
Obtain average expression from replicates
}
