% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{medianScaling}
\alias{medianScaling}
\title{Median centering and scaling}
\usage{
medianScaling(mat, scale = FALSE, grps = NULL, reorder = FALSE, assay = NULL)
}
\arguments{
\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to samples.}

\item{scale}{a boolean flag indicating whether to scale the samples.}

\item{grps}{a string or factor specifying the grouping (replciates).}

\item{reorder}{To reorder the columns by group (\code{grps}).
By default (\code{reorder=FALSE}), original column order is maintained.}

\item{assay}{an assay to be selected if \code{mat} is a PhosphoExperiment
object.}
}
\value{
A median scaled matrix
}
\description{
Median centering and scaling of an input numeric matrix
}
