% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{minmax}
\alias{minmax}
\title{Minmax scaling}
\usage{
minmax(mat)
}
\arguments{
\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to condition}
}
\value{
Minmax standardised matrix
}
\description{
Perform a minmax standardisation to scale data into 0 to 1 range
}
