% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{mIntersect}
\alias{mIntersect}
\alias{mUnion}
\title{Multi-intersection, union}
\usage{
mIntersect(x, y, ...)
mUnion(x, y, ...)
}
\arguments{
\item{x, y, ...}{objects to find intersection/union.}
}
\value{
An intersection/union of input parameters
}
\description{
A recusive loop for intersecting multiple sets.
}
