% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization_ProTN}
\alias{normalization_ProTN}
\title{Normalize Data}
\usage{
normalization_ProTN(proteome_data, spatial_data = FALSE)
}
\arguments{
\item{proteome_data}{A list containing proteome data, including peptide annotations, protein-level data, and peptide-level data.}

\item{spatial_data}{Logical. TRUE if is spatial dataset}
}
\value{
A list containing the normalized proteome data:
\item{dat_gene}{Data table containing the protein-level normalized data.}
\item{dat_pep}{Data table containing the peptide-level normalized data.}
}
\description{
This function normalizes data by summarizing peptide intensities into protein-level data, applying median-based normalization.
It also performs median centering for peptide-level log2 data.
}
\examples{
\dontrun{
proteome_data <- normalization_ProTN(proteome_data)
}

}
