% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{pathwayOverrepresent}
\alias{pathwayOverrepresent}
\title{phosphosite/Gene set over-representation analysis}
\usage{
pathwayOverrepresent(geneSet, annotation, universe, alter = "greater")
}
\arguments{
\item{geneSet}{an array of gene or phosphosite IDs (IDs are gene symbols etc
that match to your pathway annotation list).}

\item{annotation}{a list of pathways with each element containing an array of
gene or phosphosite IDs.}

\item{universe}{the universe/backgrond of all genes or phosphosites in your
profiled dataset.}

\item{alter}{test for enrichment ('greater', default), depletion ('less'), or
'two.sided'.}
}
\value{
A matrix of pathways and their associated substrates and p-values.
}
\description{
This function performes phosphosite (or gene) set
over-representation analysis using Fisher's exact test.
}
