% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{pathwayRankBasedEnrichment}
\alias{pathwayRankBasedEnrichment}
\title{Phosphosite/Gene set enrichment analysis}
\usage{
pathwayRankBasedEnrichment(geneStats, annotation, alter = "greater")
}
\arguments{
\item{geneStats}{an array of statistics (e.g. log2 FC) of all quantified
genes or phosphosite with names of the array as gene or phosphosite IDs.}

\item{annotation}{a list of pathways with each element containing an array of
gene IDs.}

\item{alter}{test for enrichment ('greater', default), depletion ('less'), or
'two.sided'.}
}
\value{
A matrix of pathways and their associated substrates and p-values.
}
\description{
This function performes phosphosite (or gene) set enrichment analysis using
Wilcoxon Rank Sum test.
}
