% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis_figure.R
\name{pca_differential_analysis_plot}
\alias{pca_differential_analysis_plot}
\title{PCA Plot}
\usage{
pca_differential_analysis_plot(differential_analysis, proteome_data, type)
}
\arguments{
\item{proteome_data}{A list containing proteome data, including sample annotation and either protein or peptide data matrices.}

\item{type}{Character; specifies the type of data to plot. Options are "protein" or "peptide".}
}
\value{
A list containing:
\item{pca_dt}{Data table containing the PCA coordinates for each sample.}
\item{plot}{ggplot2 object representing the PCA plot.}
}
\description{
This function generates a Principal Component Analysis (PCA) plot for either protein or peptide data based on the first two principal components.
It supports the inclusion of phosphoproteomics data if available and provides a variance explanation for the axes.
}
\examples{
\dontrun{
result <- pca_plot(proteome_data, type = "protein")
print(result$plot)
}

}
