% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{perform_enrichment_analysis}
\alias{perform_enrichment_analysis}
\title{Perform Enrichment Analysis on Differential Results}
\usage{
perform_enrichment_analysis(
  differential_results,
  dirOutput = "results_ProTN",
  subfold_Tab = "table",
  subfold_Dat = "rdata",
  pval_fdr_enrich = "p_adj",
  pval_enrich_thr = 0.05,
  overlap_size_enrich_thr = 5,
  enrichR_custom_DB = FALSE,
  enrich_filter_DBs = NULL,
  phospho_ctrl = FALSE
)
}
\arguments{
\item{differential_results}{A list containing differential results, either \code{protein_results_long}
for proteomics data or \code{peptide_results_long} for phospho-proteomics data.}

\item{dirOutput}{The directory where the results file will be saved. Default is "results_ProTN".}

\item{subfold_Tab}{The subfolder within \code{dirOutput} where the Excel file will be saved. Default is "table".}

\item{subfold_Dat}{The subfolder within \code{dirOutput} where the RData file will be saved. Default is "rdata".}

\item{pval_fdr_enrich}{The column name in the differential results containing the p-value after FDR correction.
Default is \code{"p_adj"}.}

\item{pval_enrich_thr}{The threshold for the p-value (adjusted) to consider significant enrichment.
Default is 0.05.}

\item{overlap_size_enrich_thr}{The threshold for the size of the overlap (number of DEPs in the term)
to consider significant enrichment. Default is 5.}

\item{enrichR_custom_DB}{A logical flag to indicate if custom EnrichR databases should be used. Default is \code{FALSE}.}

\item{enrich_filter_DBs}{A vector of custom EnrichR databases to filter enrichment results.
This argument is used if \code{enrichR_custom_DB} is set to \code{TRUE}.}

\item{phospho_ctrl}{A logical flag to indicate if phospho control data should be excluded.
Default is \code{FALSE}.}
}
\value{
A data.table containing the enrichment results, including terms, significance scores,
overlap sizes, and other relevant information.
}
\description{
This function performs enrichment analysis using the EnrichR tool based on the results
of differential expression analysis (proteomics or phospho-proteomics). The function
generates enrichment results and saves them as both an Excel file and an RData object
for further analysis.
}
\details{
It supports custom EnrichR databases and allows the user to filter enrichment results
based on p-value, FDR threshold, and overlap size.

The function generates two types of output files:
\itemize{
\item \code{enrichment.xlsx}: An Excel file containing the enrichment results.
\item \code{enrichment.RData}: A saved RData file with the filtered enrichment results for further analysis.
}
}
\examples{
# Example usage
perform_enrichment_analysis(differential_results = differential_results_example,
                            dirOutput = "results_directory")

}
