% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{phosCollapse}
\alias{phosCollapse}
\title{Summarising phosphosites to proteins}
\usage{
phosCollapse(mat, id, stat, by='min')
}
\arguments{
\item{mat}{a matrix with rows correspond to phosphosites and columns
correspond to samples.}

\item{id}{an array indicating the groupping of phosphosites etc.}

\item{stat}{an array containing statistics of phosphosite such as
phosphorylation levels.}

\item{by}{how to summarise phosphosites using their statistics. Either by
'min' (default), 'max', or 'mid'.}
}
\value{
A matrix summarised to protein level
}
\description{
Summarising phosphosite-level information to proteins for performing
downstream gene-centric analyses.
}
