% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{plot_abundance_distribution}
\alias{plot_abundance_distribution}
\title{Plot Abundance Distribution}
\usage{
plot_abundance_distribution(proteome_data, type)
}
\value{
A list containing:
\itemize{
\item \code{dt}: A data.table of log2-normalized abundance values with sample annotations
\item \code{plot}: A ggplot2 object representing the abundance distribution as a violin and boxplot.
}
}
\description{
This function generates a violin and boxplot representation of log2-normalized abundance values for proteins or peptides
}
\examples{
\dontrun{
result <- plot_abundance_distribution(proteome_data, type = "protein")
}
}
