% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{plot_selected_proteins}
\alias{plot_selected_proteins}
\title{Plot Selected Proteins}
\usage{
plot_selected_proteins(proteome_data, list_protein)
}
\arguments{
\item{proteome_data}{A list containing proteome data, including gene intensity data and sample annotations.}

\item{list_protein}{A character vector of gene names to be plotted.}
}
\value{
A list containing:
\item{dt}{Data table of protein intensity values and the corresponding sample and condition information.}
\item{plot}{ggplot2 object representing the protein abundance plot for the selected proteins.}
}
\description{
This function generates a plot for the selected proteins' abundance across different conditions. It creates a crossbar plot
with error bars representing the average intensity and standard error of selected proteins. If phosphoproteomics data is available,
it uses the corresponding data, otherwise, it uses the proteomics data.
}
\examples{
\dontrun{
result <- plot_selected_proteins(proteome_data, list_protein = c("ProteinA", "ProteinB"))
print(result$plot)
}

}
