% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{plotQC}
\alias{plotQC}
\title{A set of function for data QC plot}
\usage{
plotQC(mat, grps, labels, panel = 
c("quantify", "dendrogram", "abundance", "pca", "all"))
}
\arguments{
\item{mat}{A p by n matrix, where p is the number of phosphosites and n is
the number of samples.}

\item{grps}{A vector of colours to be used in the plot. The length should be
equal to the columns of the mat.}

\item{labels}{A vector of sample names. Used the label points in PCA plot
(panel=4)}

\item{panel}{A type of plot to output. See description for details.}
}
\value{
A graphical plot
}
\description{
The \code{panel} parameter allows different type of visualisation for output
object from PhosR.
\code{panel = "all"} is used to create a 2*2 panel of plots including the
following.
\code{panel = "quantify"} is used to visualise percentage of quantification after
imputataion.
\code{panel = "dendrogram"} is used to visualise dendrogram (hierarchical
clustering) of the input matrix.
\code{panel = "abundance"} is used to visualise abundance level of samples from
the input matrix.
\code{panel = "pca"} is used to show PCA plot
}
