% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{plotSignalomeMap}
\alias{plotSignalomeMap}
\title{Plot signalome map}
\usage{
plotSignalomeMap(signalomes, color)
}
\arguments{
\item{signalomes}{output from \code{Signalomes} function}

\item{color}{a string specifying the color vector for kinases}
}
\value{
a ggplot object
}
\description{
Plot signalome map
}
