% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_phospho_proteome_proteomics}
\alias{read_phospho_proteome_proteomics}
\title{Read Phospho-Proteome Data}
\usage{
read_phospho_proteome_proteomics(
  software,
  folder_proteome,
  folder_phospho,
  keep_only_phosphomodification = T,
  phospho_thr = 0.75,
  annotation_proteome_filename = "annotation",
  peptide_proteome_filename = "pep",
  proteinGroup_proteome_filename = "prot",
  condition_proteome_col = "Condition",
  sample_proteome_col = "Sample",
  color_proteome_col = "Color",
  annotation_phospho_filename = "annotation",
  peptide_phospho_filename = "pep",
  proteinGroup_phospho_filename = "prot",
  psm_phospho_filename = "psm",
  condition_phospho_col = "Condition",
  sample_phospho_col = "Sample",
  color_phospho_col = "Color",
  batch_corr_exe = FALSE,
  batch_col = "batch",
  filt_absent_value = 0,
  min_peptide_protein = 0
)
}
\arguments{
\item{software}{Character; the proteomics software used, either "PD" for Proteome Discoverer or "MQ" for MaxQuant.}

\item{folder_proteome}{Character; the folder containing proteome data files.}

\item{folder_phospho}{Character; the folder containing phosphoproteome data files.}

\item{keep_only_phosphomodification}{Logical; whether to keep only phospho-modified peptides. Default is TRUE.}

\item{phospho_thr}{Numeric; the threshold for phospho-modified peptides. Default is 0.75.}

\item{annotation_proteome_filename}{Character; the name or pattern of the proteome annotation file. Default is "annotation".}

\item{peptide_proteome_filename}{Character; the name or pattern of the proteome peptide file. Default is "pep".}

\item{proteinGroup_proteome_filename}{Character; the name or pattern of the proteome protein group file. Default is "prot".}

\item{condition_proteome_col}{Character; the column name representing condition information in the proteome annotation file. Default is "Condition".}

\item{sample_proteome_col}{Character; the column name representing sample information in the proteome annotation file. Default is "Sample".}

\item{color_proteome_col}{Character; the column name representing color information for the plot in the proteome annotation file. Default is "Color".}

\item{annotation_phospho_filename}{Character; the name or pattern of the phosphoproteome annotation file. Default is "annotation".}

\item{peptide_phospho_filename}{Character; the name or pattern of the phosphoproteome peptide file. Default is "pep".}

\item{proteinGroup_phospho_filename}{Character; the name or pattern of the phosphoproteome protein group file. Default is "prot".}

\item{psm_phospho_filename}{Character; the name or pattern of the phosphoproteome PSM file. Default is "psm".}

\item{condition_phospho_col}{Character; the column name representing condition information in the phosphoproteome annotation file. Default is "Condition".}

\item{sample_phospho_col}{Character; the column name representing sample information in the phosphoproteome annotation file. Default is "Sample".}

\item{color_phospho_col}{Character; the column name representing color information for the plot in the phosphoproteome annotation file. Default is "Color".}

\item{batch_corr_exe}{Logical; whether batch correction should be applied. Default is FALSE.}

\item{batch_col}{Character; the column name representing batch information in the annotation file. Default is "batch".}

\item{filt_absent_value}{Numeric; the value used to filter out absent data. Default is 0.}

\item{min_peptide_protein}{Numeric; the value used to filter out protein by N peptide. Default is 0.}
}
\value{
A list containing both proteome and phosphoproteome data, including annotation and peptide data,
with batch correction applied if specified and only phospho-modified peptides kept if requested.
}
\description{
This function reads both proteome and phosphoproteome data from files based on the specified software (Proteome Discoverer or MaxQuant).
It supports reading annotation, peptide, protein group, and PSM data for both proteome and phosphoproteome, with options for batch correction,
filtering absent values, and keeping only phospho-modified peptides above a specified threshold.
}
\examples{
\dontrun{
phospho_proteome_data <- read_phospho_proteome_proteomics(software = "MQ", 
                                                        folder_proteome = "proteome_folder", 
                                                        folder_phospho = "phospho_folder")
}

}
