% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_phosphoproteomics}
\alias{read_phosphoproteomics}
\title{Read Phosphoproteomics Data}
\usage{
read_phosphoproteomics(
  software,
  folder,
  keep_only_phosphomodification = T,
  phospho_thr = 0.75,
  annotation_filename = "annotation",
  peptide_filename = "pep",
  proteinGroup_filename = "prot",
  psm_filename = "psm",
  condition_col = "Condition",
  sample_col = "Sample",
  color_col = "Color",
  batch_corr_exe = FALSE,
  batch_col = "batch",
  filt_absent_value = 0,
  min_peptide_protein = 0
)
}
\arguments{
\item{software}{Character; the proteomics software used, either "PD" for Proteome Discoverer or "MQ" for MaxQuant.}

\item{folder}{Character; the folder containing the data files.}

\item{keep_only_phosphomodification}{Logical; whether to keep only phospho-modified peptides. Default is TRUE.}

\item{phospho_thr}{Numeric; the threshold for phospho-modified peptides. Default is 0.75.}

\item{annotation_filename}{Character; the name or pattern of the annotation file. Default is "annotation".}

\item{peptide_filename}{Character; the name or pattern of the peptide file. Default is "pep".}

\item{proteinGroup_filename}{Character; the name or pattern of the protein group file. Default is "prot".}

\item{psm_filename}{Character; the name or pattern of the PSM file. Default is "psm".}

\item{condition_col}{Character; the column name representing condition information in the annotation file. Default is "Condition".}

\item{sample_col}{Character; the column name representing sample information in the annotation file. Default is "Sample".}

\item{color_col}{Character; the column name representing color information for the plot in the annotation file. Default is "Color".}

\item{batch_corr_exe}{Logical; whether batch correction should be applied. Default is FALSE.}

\item{batch_col}{Character; the column name representing batch information in the annotation file. Default is "batch".}

\item{filt_absent_value}{Numeric; the value used to filter out absent data. Default is 0.}

\item{min_peptide_protein}{Numeric; the value used to filter out protein by N peptide. Default is 0.}
}
\value{
A list containing phosphoproteomics data, including annotation and peptide data,
with batch correction applied if specified and only phospho-modified peptides kept if requested.
}
\description{
This function reads phosphoproteomics data from files based on the specified software (Proteome Discoverer or MaxQuant).
It supports reading annotation, peptide, protein group, and PSM data, with options for batch correction, filtering absent values,
and keeping only phospho-modified peptides above a specified threshold.
}
\examples{
\dontrun{
proteome_data <- read_phosphoproteomics(software = "MQ", folder = "data_folder")
}

}
