% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_proteomics}
\alias{read_proteomics}
\title{Read Proteomics Data}
\usage{
read_proteomics(
  software,
  folder,
  use_proteinGroups_MQ = FALSE,
  annotation_filename = "annotation",
  peptide_filename = "pep",
  proteinGroup_filename = "prot",
  condition_col = "Condition",
  sample_col = "Sample",
  color_col = "Color",
  batch_corr_exe = FALSE,
  batch_col = "batch",
  filt_absent_value = 0,
  min_peptide_protein = 0
)
}
\arguments{
\item{software}{Character; the proteomics software used, either "PD" for Proteome Discoverer or "MQ" for MaxQuant.}

\item{folder}{Character; the folder containing the data files.}

\item{use_proteinGroups_MQ}{Logical; Only for MaxQuant. If FALSE it expect the evidence.txt file, if TRUE require peptide.txt and proteinGroups.txt files. Default is FALSE-}

\item{annotation_filename}{Character; the name or pattern of the annotation file. Default is "annotation".}

\item{peptide_filename}{Character; the name or pattern of the peptide file. Default is "pep".}

\item{proteinGroup_filename}{Character; the name or pattern of the protein group file. Default is "prot".}

\item{condition_col}{Character; the column name representing condition information in the annotation file. Default is "Condition".}

\item{sample_col}{Character; the column name representing sample information in the annotation file. Default is "Sample".}

\item{color_col}{Character; the column name representing color information for the plot in the annotation file. Default is "Color".}

\item{batch_corr_exe}{Logical; whether batch correction should be applied. Default is FALSE.}

\item{batch_col}{Character; the column name representing batch information in the annotation file. Default is "batch".}

\item{filt_absent_value}{Numeric; the value used to filter out absent data. Default is 0.}

\item{min_peptide_protein}{Numeric; the value used to filter out protein by N peptide. Default is 0.}
}
\value{
A list containing proteomics data, including annotation and peptide data,
with batch correction applied if specified.
}
\description{
This function reads proteomics data from files based on the specified software (Proteome Discoverer or MaxQuant).
It supports reading annotation, peptide, and protein group data, with options for batch correction and filtering absent values.
}
\examples{
\dontrun{
proteome_data <- read_proteomics(software = "MQ", folder = "data_folder")
}

}
