% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{save_abundance_tables}
\alias{save_abundance_tables}
\title{Save Abundance Tables for Proteomics Data}
\usage{
save_abundance_tables(
  proteome_data,
  dirOutput = "results_ProTN",
  subfolder = "table"
)
}
\arguments{
\item{proteome_data}{List; a list containing proteomics or phosphoproteomics data.}

\item{dirOutput}{Character; the directory path where the results should be saved. Default is "results_ProTN".}

\item{subfolder}{Character; a subfolder within \code{dirOutput} where the results will be stored. Default is "table".}
}
\description{
This function processes proteomics and phosphoproteomics data to compute summary statistics and save the results in an Excel file. It includes normalized, imputed, and batch-corrected abundance values for proteins and peptides, per sample and per condition.
}
