% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{save_differential_analysis_table}
\alias{save_differential_analysis_table}
\title{Save Differential Analysis Results to Excel}
\usage{
save_differential_analysis_table(
  proteome_data,
  differential_results,
  dirOutput = "results_ProTN",
  subfolder = "table",
  phospho_ctrl = FALSE,
  phospho_analysis = FALSE
)
}
\arguments{
\item{proteome_data}{A list containing proteomic data.}

\item{differential_results}{A list containing differential expression results.}

\item{dirOutput}{The directory where the results file will be saved. Default is "results_ProTN".}

\item{subfolder}{The subfolder within \code{dirOutput} where the file will be saved. Default is "table".}

\item{phospho_ctrl}{A logical flag to indicate if phospho control data should be excluded.
Default is \code{FALSE}.}

\item{phospho_analysis}{A logical flag to indicate if is a phosphorylation analysis.
Default is \code{FALSE}.}
}
\description{
This function saves the results of a differential analysis (proteomics or phospho-proteomics)
into an Excel file. It merges the results with the sample annotation data and generates a
comprehensive Excel file with the analysis results, including protein and peptide-level
differential expression data.
}
\examples{
\dontrun{
save_differential_analysis_table(proteome_data = proteome_data_example,
                                 differential_results = differential_results_example,
                                 dirOutput = "results_directory")
}
}
