% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{scImpute}
\alias{scImpute}
\title{Site- and condition-specific (sc) impute}
\usage{
scImpute(mat, percent, grps, assay)
}
\arguments{
\item{mat}{a matrix (or PhosphoExperiment object) with rows correspond to
phosphosites and columns correspond to replicates within a condition.}

\item{percent}{a percent from 0 to 1, specifying the percentage of quantified
values in any treatment group.}

\item{grps}{a string specifying the grouping (replciates).}

\item{assay}{an assay to be selected if \code{mat} is a PhosphoExperiment
object.}
}
\value{
An imputed matrix. If param \code{mat} is a PhosphoExperiment
object, a PhosphoExperiment object will be returned.
}
\description{
Impute the missing values for a phosphosite across replicates within a single
condition (or treatment)
if there are n or more quantified values of that phosphosite in that
condition.
}
