% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{selectLocalisedSites}
\alias{selectLocalisedSites}
\title{Select phosphosites by localisation score}
\usage{
selectLocalisedSites(mat, loc=NULL, prob = 0.75)
}
\arguments{
\item{mat}{a matrix (or PhosphoExperiment object) with rows corresponding to
phosphosites and columns corresponding to samples in replicates for different
treatments.}

\item{loc}{a vector of localisation scores}

\item{prob}{a percent from 0 to 1, specifying the localisation probability
of quantified values in across all samples for retaining a phosphosite for
subsequent analysis.}
}
\value{
a filtered matrix
}
\description{
Select phosphosites with a localisation score higher than the
pre-defined probability score (default score = 0.75)
}
