% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{selectTimes}
\alias{selectTimes}
\title{selectTimes}
\usage{
selectTimes(mat, timepoint, order, percent, w, assay)
}
\arguments{
\item{mat}{a matrix (or PhosphoExperiment object) with rows correspond to
phosphosites and columns correspond to samples in replicates for different
treatments.}

\item{timepoint}{a timepoint as factor with a length equal to the
number of columns of mat.}

\item{order}{a vector specifying the order of timepoints.}

\item{percent}{a percent (decimal) from 0 to 1, to filter phosphosites with
with missing value larger than percent per timepoint.}

\item{w}{a timepoint window for selection of phosphosites to remove.}

\item{assay}{an assay to be selected if \code{mat} is a PhosphoExperiment
object.}
}
\value{
a filtered matrix. If param \code{mat} is a SummarizedExperiment
object, a SummarizedExperiment object will be returned.
}
\description{
selectTimes
}
