% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichR.R
\name{set_enrichR}
\alias{set_enrichR}
\title{onLoad hook to setup package options}
\usage{
set_enrichR()
}
\description{
onLoad hook to setup package options
}
\details{
onLoad hook to setup package options and to check connection to website
}
\author{
Wajid Jawaid \email{wajid.jawaid@gmail.com}
}
