% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{Signalomes}
\alias{Signalomes}
\title{PhosR Signalomes}
\usage{
Signalomes(KSR, predMatrix, exprsMat, KOI, threskinaseNetwork=0.9,
signalomeCutoff=0.5, module_res = NULL, filter = FALSE, verbose = TRUE)
}
\arguments{
\item{KSR}{kinase-substrate relationship scoring results}

\item{predMatrix}{output of kinaseSubstratePred function}

\item{exprsMat}{a matrix with rows corresponding to phosphosites and columns
corresponding to samples}

\item{KOI}{a character vector that contains kinases of interest for which
expanded signalomes will be generated}

\item{threskinaseNetwork}{threshold used to select interconnected kinases for
the expanded signalomes}

\item{signalomeCutoff}{threshold used to filter kinase-substrate
relationships}

\item{module_res}{parameter to select number of final modules}

\item{filter}{parameter to filter modules with only few proteins}

\item{verbose}{Default to \code{TRUE} to show messages during the progress.
All messages will be suppressed if set to \code{FALSE}}
}
\value{
A list of 3 elements.
\code{Signalomes}, \code{proteinModules} and \code{kinaseSubstrates}
}
\description{
A function to generate signalomes
}
