% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{siteAnnotate}
\alias{siteAnnotate}
\title{Phosphosite annotation}
\usage{
siteAnnotate(site, phosScoringMatrices, predMatrix)
}
\arguments{
\item{site}{site the ID of a phosphosite}

\item{phosScoringMatrices}{output from function kinaseSubstrateScore_local()}

\item{predMatrix}{a prediction matrix from kinaseSubstratePred()}
}
\value{
A graphical plot
}
\description{
This function plots the combined scores of each of all kinases
for a given phosphosites
}
