% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosR.R
\name{standardise}
\alias{standardise}
\title{Standardisation}
\usage{
standardise(mat)
}
\arguments{
\item{mat}{a matrix (or a PhosphoExperiment object) with rows correspond
to phosphosites and columns correspond to samples.}
}
\value{
A standardised matrix
}
\description{
Standardisation by z-score transformation.
}
