% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRING_analysis.R
\name{STRINGdb_network}
\alias{STRINGdb_network}
\title{STRINGdb Network Analysis}
\usage{
STRINGdb_network(
  differential_results,
  species = "Homo sapiens",
  dirOutput = "results_ProTN",
  subfold_Fig = "pics",
  subfold_net = "STRINGdb",
  phospho_ctrl = FALSE,
  score_thr = 700,
  shiny = FALSE
)
}
\arguments{
\item{differential_results}{A list containing differential expression results, either at the protein
or peptide level. It should include \code{protein_results_long} or \code{peptide_results_long}.}

\item{species}{Character. The species name (default: \code{"Homo sapiens"}).}

\item{dirOutput}{Character. The main output directory (default: \code{"results_ProTN"}).}

\item{subfold_Fig}{Character. Subfolder for figures (default: \code{"pics"}).}

\item{subfold_net}{Character. Subfolder for STRINGdb network results (default: \code{"STRINGdb"}).}

\item{phospho_ctrl}{Logical. Whether to include phospho-control comparisons (default: \code{FALSE}).}

\item{score_thr}{Numeric. Threshold score for STRINGdb edge. (default: \code{700})}

\item{shiny}{Logical. Return results for the shiny app (default: \code{FALSE})}
}
\value{
A data frame with STRINGdb network results.
}
\description{
This function performs a STRINGdb network analysis on differentially expressed proteins or peptides.
It processes input differential expression results and evaluates STRING network interactions.
}
\examples{
\dontrun{
  results <- STRINGdb_network(differential_results, species="Mus musculus")
}
}
