% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_figure.R
\name{complexity_plot}
\alias{complexity_plot}
\title{Complexity Plot}
\usage{
complexity_plot(
  proteome_data,
  phospho_with_proteome = FALSE,
  col_vec = c("#008752", "#8FD4B8", "#7AC2E0", "#5BA4D0")
)
}
\arguments{
\item{proteome_data}{A list containing \code{c_anno} or \code{c_anno_phospho}, \code{psm_log_prot_df}, and \code{psm_log_pet_df}.}

\item{phospho_with_proteome}{Logical. Whether to include phosphoproteome data alongside proteome data (default: \code{FALSE}).}

\item{col_vec}{Vector. Vector of 4 color. Default: "#008752", "#8FD4B8", "#7AC2E0", "#5BA4D0"}
}
\value{
A list containing:
\itemize{
\item \code{dt}: Data table for abundance.
\item \code{plot}: A ggplot2 object visualizing abundance.
}
}
\description{
This function generates a complexity plot of raw abundance. Provide the data before normalization and imputation
}
\examples{
\dontrun{
  abundance_plot <- complexity_plot(proteome_data)
}
}
